(y60) fixed signature of virtual functions in viewacm

+++ y60/products/ViewACM/MovieScreen.cpp
@@ -190,7 +190,7 @@ namespace video {
     }
 
     bool
-    MovieScreen::updateToFrame(unsigned long nextFrame) {
+    MovieScreen::updateToFrame(unsigned int nextFrame) {
             return update();
     }
